+++ StockExchangeSim.Core/Models/SampleOrderDetail.cs
@@ -13,17 +13,17 @@ namespace StockExchangeSim.Core.Models
 
         public int Quantity { get; set; }
 
-        public double Discount { get; set; }
+        public float Discount { get; set; }
 
         public string QuantityPerUnit { get; set; }
 
-        public double UnitPrice { get; set; }
+        public float UnitPrice { get; set; }
 
         public string CategoryName { get; set; }
 
         public string CategoryDescription { get; set; }
 
-        public double Total { get; set; }
+        public float Total { get; set; }
 
         public string ShortDescription => $"Product ID: {ProductID} - {ProductName}";
     }
